update tests for clearer test cases.
